+++ docx-wasm/js/json/bindings/FooterReference.ts
@@ -1,0 +1,2 @@
+
+export interface FooterReference { footerType: string, id: string, }