+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## @0.4.8 (19. Feb, 2024)
+
+- Fixed a bug, image in header/footer is not stored in media when read.
+- Fixed a bug, image in header/footer is broken.
+
 ## docx-wasm@0.0.278-rc27 (17. Jan, 2024)
 
 - Support part of `pPrDefault`.